Move link and update interface types to the nodes types.
