+++ ui/src/app/machines/views/MachineDetails/MachineNetwork/EditBondForm/EditBondForm.tsx
@@ -29,10 +29,7 @@ import {
 } from "app/store/general/types";
 import { actions as machineActions } from "app/store/machine";
 import machineSelectors from "app/store/machine/selectors";
-import type {
-  MachineDetails,
-  UpdateInterfaceParams,
-} from "app/store/machine/types";
+import type { MachineDetails } from "app/store/machine/types";
 import type { MachineEventErrors } from "app/store/machine/types/base";
 import {
   getInterfaceIPAddress,
@@ -44,7 +41,11 @@ import {
 import type { RootState } from "app/store/root/types";
 import { actions as subnetActions } from "app/store/subnet";
 import subnetSelectors from "app/store/subnet/selectors";
-import type { NetworkInterface, NetworkLink } from "app/store/types/node";
+import type {
+  NetworkInterface,
+  NetworkLink,
+  UpdateInterfaceParams,
+} from "app/store/types/node";
 import { actions as vlanActions } from "app/store/vlan";
 import vlanSelectors from "app/store/vlan/selectors";
 import { arrayItemsEqual } from "app/utils";
