+++ ui/src/app/store/machine/types/actions.ts
@@ -7,6 +7,7 @@ import type { Script } from "app/store/script/types";
 import type { Subnet } from "app/store/subnet/types";
 import type { NetworkLinkMode } from "app/store/types/enum";
 import type {
+  LinkParams,
   NetworkInterface,
   NetworkInterfaceParams,
   NetworkLink,
@@ -230,14 +231,6 @@ export type GetSummaryYamlParams = {
   fileId: string;
 };
 
-// Common params for methods that can accept a link.
-export type LinkParams = {
-  default_gateway?: boolean;
-  ip_address?: NetworkLink["ip_address"];
-  mode?: NetworkLinkMode;
-  subnet?: Subnet["id"];
-};
-
 export type LinkSubnetParams = {
   interface_id: NetworkInterface["id"];
   ip_address?: NetworkLink["ip_address"];
@@ -329,32 +322,6 @@ export type UpdateFilesystemParams = {
   systemId: Machine[MachineMeta.PK];
   tags?: string[];
 } & OptionalFilesystemParams;
-
-export type UpdateInterfaceParams = {
-  bridge_fd?: NetworkInterfaceParams["bridge_fd"];
-  bridge_stp?: NetworkInterfaceParams["bridge_stp"];
-  bond_downdelay?: NetworkInterfaceParams["bond_downdelay"];
-  bond_lacp_rate?: NetworkInterfaceParams["bond_lacp_rate"];
-  bond_miimon?: NetworkInterfaceParams["bond_miimon"];
-  bond_mode?: NetworkInterfaceParams["bond_mode"];
-  bond_num_grat_arp?: NetworkInterfaceParams["bond_num_grat_arp"];
-  bond_updelay?: NetworkInterfaceParams["bond_updelay"];
-  bond_xmit_hash_policy?: NetworkInterfaceParams["bond_xmit_hash_policy"];
-  bridge_type?: NetworkInterfaceParams["bridge_type"];
-  enabled?: NetworkInterface["enabled"];
-  interface_id: NetworkInterface["id"];
-  interface_speed?: NetworkInterface["interface_speed"];
-  link_connected?: NetworkInterface["link_connected"];
-  link_id?: NetworkLink["id"];
-  link_speed?: NetworkInterface["link_speed"];
-  mac_address?: NetworkInterface["mac_address"];
-  name?: NetworkInterface["name"];
-  numa_node?: NetworkInterface["numa_node"];
-  parents?: NetworkInterface["parents"];
-  system_id: Machine[MachineMeta.PK];
-  tags?: NetworkInterface["tags"];
-  vlan?: NetworkInterface["vlan_id"];
-} & LinkParams;
 
 export type UpdateParams = CreateParams & {
   [MachineMeta.PK]: Machine[MachineMeta.PK];
