+++ ui/src/app/store/machine/types/index.ts
@@ -37,7 +37,6 @@ export type {
   UnmountSpecialParams,
   UpdateDiskParams,
   UpdateFilesystemParams,
-  UpdateInterfaceParams,
   UpdateParams,
   UpdateVmfsDatastoreParams,
 } from "./actions";
